+++ 2-1_ObjectOrientedProgramming/JWC/Assignment2/Add.java
@@ -18,5 +18,7 @@ public class Add {
     }
 
     System.out.println(sum);
+
+    sc.close();
   }
 }
