+++ 2-1_ObjectOrientedProgramming/JWC/Assignment2/Question10.java
@@ -1,33 +1,44 @@
-import java.lang.Math;
-import java.util.Scanner;
+import java.util.Random;
 
 public class Question10 {
   public static void main(String[] args) {
-    Scanner scanner = new Scanner(System.in);
+    int[] temp1 = new int[16];
+    int[] temp2 = new int[6];
+    int[][] ary = new int[4][4];
+    Random rn = new Random();
 
-    int[] ary = new int[6];
-
-    System.out.print("첫 번째 원의 중심과 반지름 입력 >> ");
-    for (int i = 0; i < 3; ++i) {
-      ary[i] = scanner.nextInt();
+    for (int i = 0; i < 16; ++i) {
+      temp1[i] = rn.nextInt(10) + 1;
     }
 
-    System.out.print("두 번째 원의 중심과 반지름 입력 >> ");
-    for (int i = 3; i < 6; ++i) {
-      ary[i] = scanner.nextInt();
+    for (int i = 0; i < 6; ++i) {
+      temp2[i] = rn.nextInt(16);
+
+      for (int j = 0; j < i; ++j) {
+        if (temp2[i] == temp2[j]) {
+          --i;
+          break;
+        }
+      }
     }
 
-    // 두 원 사이 거리 구하기
-    double temp1 = Math.pow((ary[3] - ary[0]), 2);
-    double temp2 = Math.pow((ary[4] - ary[1]), 2);
-    double distance = Math.sqrt(temp1 + temp2);
+    for (int i = 0; i < 6; ++i) {
+      temp1[temp2[i]] = 0;
+    }
 
-    // 두 반지름의 합이 두 원 사이 거리보다 크면 충돌함.
-    if (distance < ary[2] + ary[5])
-      System.out.println("두 원은 서로 겹친다.");
-    else
-      System.out.println("두 원은 서로 겹치지 않는다.");
+    int idx = 0;
+    for (int i = 0; i < 4; ++i) {
+      for (int j = 0; j < 4; ++j) {
+        ary[i][j] = temp1[idx++];
+      }
+    }
 
-    scanner.close();
+    for (int i = 0; i < 4; ++i) {
+
+      for (int j = 0; j < 4; ++j) {
+        System.out.print(ary[i][j] + " ");
+      }
+      System.out.println("");
+    }
   }
 }
